SkillSync-v2--main/backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
dj-database-url>=2.1
psycopg2-binary>=2.9
python-dotenv>=1.0
gunicorn>=21.2
whitenoise>=6.6
scikit-learn>=1.3
nltk>=3.8
pandas>=2.0
numpy>=1.24
